Move Archiver to directory.
